Intel compiler warning fixes

+++ Source/Modules/Core/Game/Tile.h
@@ -44,7 +44,7 @@ struct Tile
 
     /** Constructor */
     Tile() :
-        state (blank)
+        state ((int) blank)
     {
     }
